// AutoVideoMetaLocalize/client/generated-sources/openapi/models/Thumbnail.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Meta Localize
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Thumbnail
 */
export interface Thumbnail {
    /**
     * 
     * @type {number}
     * @memberof Thumbnail
     */
    height?: number | null;
    /**
     * 
     * @type {string}
     * @memberof Thumbnail
     */
    url?: string | null;
    /**
     * 
     * @type {number}
     * @memberof Thumbnail
     */
    width?: number | null;
    /**
     * 
     * @type {string}
     * @memberof Thumbnail
     */
    eTag?: string | null;
}

export function ThumbnailFromJSON(json: any): Thumbnail {
    return ThumbnailFromJSONTyped(json, false);
}

export function ThumbnailFromJSONTyped(json: any, ignoreDiscriminator: boolean): Thumbnail {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'height': !exists(json, 'height') ? undefined : json['height'],
        'url': !exists(json, 'url') ? undefined : json['url'],
        'width': !exists(json, 'width') ? undefined : json['width'],
        'eTag': !exists(json, 'eTag') ? undefined : json['eTag'],
    };
}

export function ThumbnailToJSON(value?: Thumbnail | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'height': value.height,
        'url': value.url,
        'width': value.width,
        'eTag': value.eTag,
    };
}